DRY: Organização de código comum num controller pai - Outro controller será criado e utilizará este código comum
